fix: move logging setup to separate function
