+++ packages/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2021"
 alloy                               = { workspace = true }
 serde                               = { workspace = true }
 valence-clearing-queue-supervaults  = { workspace = true }
-log                                 = { workspace = true }
 valence-domain-clients              = { workspace = true }
 valence-authorization-utils         = { workspace = true }
 cosmwasm-std                        = { workspace = true }
@@ -15,3 +14,10 @@ valence-processor-utils             = { workspace = true }
 async-trait                         = { workspace = true }
 valence-lending-utils               = { workspace = true }
 mmvault                             = { workspace = true }
+log                                 = { workspace = true }
+env_logger                          = { workspace = true }
+multi_log                           = { workspace = true }
+opentelemetry_sdk                   = { workspace = true }
+opentelemetry-otlp                  = { workspace = true }
+opentelemetry-stdout                = { workspace = true }
+opentelemetry-appender-log          = { workspace = true }
