+++ packages/src/utils/mod.rs
@@ -2,3 +2,4 @@ pub mod mars;
 pub mod obligation;
 pub mod supervaults;
 pub mod valence_core;
+pub mod logging;
